+++ README.md
@@ -44,14 +44,14 @@
 We provide code for some typical use cases of accessing the Smart Sensor Cloud Interface.
 You can run each script by executing
 
-    python use_cases/<use_case_file_name>
+    python <use_case_nr>/<use_case_file_name>
 
-| Name                                    | Explanation                                      | Links                                                                                                                         |
-|-----------------------------------------|--------------------------------------------------|----------------------------------------------------------------------|
-| use_case_1/plant_data.py                | Show all the information about a plant           | [code](use_case_1/plant_Data.py), [usage](use_case_1)                |
-| use_case_1/list_of_assets.py            | Show all assets inside an organization           | [code](use_case_1/list_of_assets.py), [usage](use_case_1)            |
-| use_case_1/detailed_asset_data.py       | Show detailed information about an asset         | [code](use_case_1/detailed_asset_data.py), [usage](use_case_1)       |
-| use_case_2/add_notification_channel.py  | Add a notification channel to a sensor           | [code](use_case_2/add_notification_channel.py), [usage](use_case_2)  |
-| TODO                                    | TODO                                             | [code](use_case_3/TODO), [usage](use_case_3)                         |
-| use_case_4/conditon_indexes.py          | Show condition indexes of one or multiple assets | [code](use_case_4/condition_indexes.py), [usage](use_cases_4)        |
-| use_case_5/measurement_values.py        | Plot measurement values of a defined asset       | [code](use_case_5/measurement_values.py), [usage](use_case_5)        |
+| Name                                             | Explanation                                      | Links                                                                                                                         |
+|--------------------------------------------------|--------------------------------------------------|-------------------------------------------------------------------------------|
+| use_case_1/get_plant_data.py                     | Show all the information about a plant           | [code](use_case_1/get_plant_data.py), [usage](use_case_1)                     |
+| use_case_1/get_simplified_asset_data.py          | Show all assets inside an organization           | [code](use_case_1/get_simplified_asset_data.py), [usage](use_case_1)          |
+| use_case_1/get_detailed_sensor_and_asset_data.py | Show detailed information about an asset         | [code](use_case_1/get_detailed_sensor_and_asset_data.py), [usage](use_case_1) |
+| use_case_2/add_notification_channel.py           | Add a notification channel to a sensor           | [code](use_case_2/add_notification_channel.py), [usage](use_case_2)           |
+| TODO                                             | TODO                                             | [code](use_case_3/TODO), [usage](use_case_3)                                  |
+| use_case_4/get_condition_indexes.py              | Show condition indexes of one or multiple assets | [code](use_case_4/get_condition_indexes.py), [usage](use_cases_4)             |
+| use_case_5/get_historic_measurements.py          | Plot measurement values of a defined asset       | [code](use_case_5/get_historic_measurements.py), [usage](use_case_5)          |
