+++ use_case_1/get_simplified_asset_data.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-"""Example 2
+"""Use case 1: request asset details - simplified asset data
 
-This code demonstrates how to get the information of every plant in the organization
+This code demonstrates how to get a list of all the assets for each plant in the organization
 
 Example:
-    $ python plant_data.py
+    $ python get_simplified_asset_data.py
 
 """
 
@@ -34,7 +34,17 @@ def run_task(settings_file=DEFAULT_SETTINGS_FILE, debug: bool = False) -> bool:
 
     # Iterate the plant list and print all assets therein
     for plant in plants:
-        pprint(plant)
+        print('Plant {}, {}:'.format(plant['plantID'], plant['plantName']))
+        print('Assets:')
+
+        # Get list of assets
+        assets = client.get_asset_list(organization_id=client.organization_id, plant_id=plant['plantID'])
+        if len(assets) == 0:
+            print('No assets in this plant')
+        else:
+            for asset in assets:
+                pprint(asset)
+
         print()
 
     return True
@@ -42,7 +52,7 @@ def run_task(settings_file=DEFAULT_SETTINGS_FILE, debug: bool = False) -> bool:
 
 # Main body
 if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Get the information of every plant in the organization')
+    parser = argparse.ArgumentParser(description='Get information of all the assets belonging to the users current organization')
     parser.add_argument('-d', '--debug', action='store_true', help='print debug information such as the sent curl request')
     args = parser.parse_args()
